# Convert stream to list for constrained boarding search [changelog skip]