### Summary

For data sets which have a huge number of constrained transfers (see #3950), the `Stream` in `ConstrainedBoarding` search is actually a bottleneck.

# Before
```
10:20:59.409 DEBUG (DebugTimingAggregator.java:232) ┌  Routing initialization           :     0 ms
10:20:59.409 DEBUG (DebugTimingAggregator.java:232) ├  Direct street routing            :     2 ms
10:20:59.409 DEBUG (DebugTimingAggregator.java:232) ├  Direct flex routing              :     0 ms
10:20:59.409 DEBUG (DebugTimingAggregator.java:232) │┌ Creating raptor data model       :    88 ms
10:20:59.409 DEBUG (DebugTimingAggregator.java:232) │├ Access routing (325 accesses)    :    31 ms
10:20:59.409 DEBUG (DebugTimingAggregator.java:232) │├ Egress routing (91 egresses)     :     7 ms
10:20:59.409 DEBUG (DebugTimingAggregator.java:232) ││ Access/Egress routing            :    31 ms
10:20:59.409 DEBUG (DebugTimingAggregator.java:232) │├ Main routing                     :  5357 ms
10:20:59.409 DEBUG (DebugTimingAggregator.java:232) │├ Creating itineraries             :    13 ms
10:20:59.409 DEBUG (DebugTimingAggregator.java:232) ├┴ Transit routing total            :  5490 ms
10:20:59.409 DEBUG (DebugTimingAggregator.java:232) │  Routing total:                   :  5491 ms
10:20:59.410 DEBUG (DebugTimingAggregator.java:232) ├  Filtering itineraries            :     0 ms
10:20:59.410 DEBUG (DebugTimingAggregator.java:232) ├  Converting model objects         :     2 ms
10:20:59.410 DEBUG (DebugTimingAggregator.java:232) ┴  Request total                    :  5493 ms
```

# After

```
10:04:42.305 DEBUG (DebugTimingAggregator.java:232) ┌  Routing initialization           :     0 ms
10:04:42.305 DEBUG (DebugTimingAggregator.java:232) ├  Direct street routing            :     2 ms
10:04:42.306 DEBUG (DebugTimingAggregator.java:232) ├  Direct flex routing              :     0 ms
10:04:42.306 DEBUG (DebugTimingAggregator.java:232) │┌ Creating raptor data model       :    81 ms
10:04:42.306 DEBUG (DebugTimingAggregator.java:232) │├ Access routing (325 accesses)    :    23 ms
10:04:42.306 DEBUG (DebugTimingAggregator.java:232) │├ Egress routing (91 egresses)     :     7 ms
10:04:42.306 DEBUG (DebugTimingAggregator.java:232) ││ Access/Egress routing            :    23 ms
10:04:42.306 DEBUG (DebugTimingAggregator.java:232) │├ Main routing                     :  3670 ms
10:04:42.306 DEBUG (DebugTimingAggregator.java:232) │├ Creating itineraries             :    17 ms
10:04:42.306 DEBUG (DebugTimingAggregator.java:232) ├┴ Transit routing total            :  3793 ms
10:04:42.306 DEBUG (DebugTimingAggregator.java:232) │  Routing total:                   :  3793 ms
10:04:42.306 DEBUG (DebugTimingAggregator.java:232) ├  Filtering itineraries            :     0 ms
10:04:42.306 DEBUG (DebugTimingAggregator.java:232) ├  Converting model objects         :     2 ms
10:04:42.306 DEBUG (DebugTimingAggregator.java:232) ┴  Request total                    :  3796 ms
```

### Issue

Relates to #3950 but I feel that in order to solve the problem we need to tackle it algorithmically.

### Unit tests

None, since it's a performance improvement.

### Code style
yes

### Documentation
no